+++ db/migrate/20210313201739_add_pack_size_to_pharmacy_batch_item.rb
@@ -1,0 +1,5 @@
+class AddPackSizeToPharmacyBatchItem < ActiveRecord::Migration[5.2]
+  def change
+    add_column :pharmacy_batch_items, :pack_size, :integer, null: true
+  end
+end